requirements.txt
streamlit
pandas
requests
feedparser
python-dotenv
pytrends
plotly
beautifulsoup4
google-generativeai
openai
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client